feat: imitation refetch fn
