+++ packages/client/src/features/topic/topic-form/topic-form.ui.tsx
@@ -1,16 +1,14 @@
-import { createTopic } from '@/entities/topic/topic.api'
 import { TopicSchemaData } from '@/entities/topic/topic.types'
 import { FormInput } from '@/shared/ui/form-input/form-input.ui'
 import { Button } from 'antd'
 import { useFormContext } from 'react-hook-form'
 
-export const TopicForm = () => {
-  const { handleSubmit, reset, control } = useFormContext<TopicSchemaData>()
+interface TopicFormProps {
+  onSubmit: (data: TopicSchemaData) => void
+}
 
-  const onSubmit = async (data: TopicSchemaData) => {
-    await createTopic(data)
-    reset()
-  }
+export const TopicForm = ({ onSubmit }: TopicFormProps) => {
+  const { handleSubmit, control } = useFormContext<TopicSchemaData>()
 
   return (
     <form onSubmit={handleSubmit(onSubmit)} className="space-y-4 mb-6">
