+++ packages/client/src/widgets/forum/forum-feed.ui.tsx
@@ -1,3 +1,4 @@
+import { createTopic, getTopics } from '@/entities/topic/topic.api'
 import { TopicSchema } from '@/entities/topic/topic.contract'
 import {
   StyledForumPageContainer,
@@ -6,7 +7,9 @@ import {
 import { TopicSchemaData } from '@/entities/topic/topic.types'
 import { TopicForm } from '@/features/topic/topic-form/topic-form.ui'
 import { TopicList } from '@/features/topic/topic-list/topic-list.ui'
+import { Topic } from '@/shared/types/Topic'
 import { zodResolver } from '@hookform/resolvers/zod'
+import { useCallback, useEffect, useState } from 'react'
 import { FormProvider, useForm } from 'react-hook-form'
 
 export const ForumWidget = () => {
@@ -15,12 +18,36 @@ export const ForumWidget = () => {
     resolver: zodResolver(TopicSchema),
   })
 
+  const [topics, setTopics] = useState<Topic[]>([])
+  const [isLoading, setIsLoading] = useState(true)
+
+  const fetchTopics = useCallback(async () => {
+    try {
+      const data = await getTopics()
+      setTopics(data)
+    } catch (error) {
+      console.error('Ошибка при получении тем:', error)
+    } finally {
+      setIsLoading(false)
+    }
+  }, [])
+
+  useEffect(() => {
+    fetchTopics()
+  }, [fetchTopics])
+
+  const onSubmit = async (data: TopicSchemaData) => {
+    await createTopic(data)
+    await fetchTopics()
+    methods.reset()
+  }
+
   return (
     <FormProvider {...methods}>
       <StyledForumPageContainer>
         <StyledForumTitle level={2}>Форум</StyledForumTitle>
-        <TopicForm />
-        <TopicList />
+        <TopicForm onSubmit={onSubmit} />
+        <TopicList topics={topics} isLoading={isLoading} />
         {/* <ForumCard /> */}
       </StyledForumPageContainer>
     </FormProvider>
